feat(background App): Updated background to set light and dark mode
The following implementation updates the app to create a light and dark mode when user switches elements
