+++ client/src/Components/Layout/Header.tsx
@@ -1,10 +1,19 @@
-import { AppBar, Toolbar, Typography } from "@mui/material";
+import { AppBar, Toolbar, Typography, Switch } from "@mui/material";
 
-const Header = () => {
+interface Props {
+  darkMode: boolean;
+  handleThemeChange: () => void;
+}
+
+const Header = ({ darkMode, handleThemeChange }: Props) => {
   return (
     <AppBar position="static" sx={{ mb: 4 }}>
       <Toolbar>
         <Typography variant="h6">Ecommerce Store</Typography>
+        <Switch
+          checked={darkMode}
+          onChange={handleThemeChange}
+        />
       </Toolbar>
     </AppBar>
   );
